fix test

+++ src/platform_cfg.cpp
@@ -351,7 +351,7 @@ void operator >> (const YAML::Node& node, CMacYamlInfo & mac_info) {
             } else {
                 it.first() >> label;
                 
-                mac_info.m_mpls_ethtype[mac_info.m_mpls_count] = {label, ethtype};
+                mac_info.m_mpls_ethtype[mac_info.m_mpls_count] = {label, (uint16_t)ethtype};
                 mac_info.m_mpls_count += 1;
             }
         }
